requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic[email]
